// src/Interfaces/LoggingTypes.ts
export enum LoggingTypes {
  alert = "alert",
  warning = "warning",
  info = "info",
  debug = "debug",
}

export enum ComponentEnums {
  GAMESESSION = "GAMESESSION",
  ADDSPELL = "ADDSPELL",
  SOCKETRECEIVER = "SOCKETRECEIVER",
  SPELLRECORD = "SPELLRECORD",
  SPELLSTATE = "SPELLSTATE",
  ADDINITIATIVE = "ADDINITIATIVE",
  EFFECT = "EFFECT",
  EFFECTCONTAINER = "EFFECTCONTAINER",
  INITIATIVEHEADER = "INITIATIVEHEADER",
  INITIATIVESTATE = "INITIATIVESTATE",
  INITRECORD = "INITRECORD",
  SORTABLELIST = "SORTABLELIST",
  INPUTEDITOR = "INPUTEDITOR",
}

export enum StoreEnums {
  updateId = "updateId",
  getInitial = "getInitial",
  getInitialSpells = "getInitialSpells",
  updateCharacterItem = "updateCharacterItem",
  updateCharacterRecord = "updateCharacterRecord",
  addCharacter = "addCharacter",
  removeCharacter = "removeCharacter",
  getSpells = "getSpells",
  addSpell = "addSpell",
  onDrop = "onDrop",
  updateSpell = "updateSpell",
  updateSpellItem = "updateSpellItem",
  changeAllCharacterStatus = "changeAllCharacterStatus",
  changeOneCharacterStatus = "changeOneCharacterStatus",
  roundStart = "roundStart",
  getInitiative = "getInitiative",
  getSorted = "getSorted",
  reSort = "reSort",
  setCurrent = "setCurrent",
  nextTurn = "nextTurn",
  previousTurn = "previousTurn",
  toDiscord = "toDiscord",
  roomSetup = "roomSetup",
  updateAll = "updateAll",
  removeSpell = "removeSpell",
  spellsDoubleArray = "spellsDoubleArray",
  resetAll = "resetAll",
  reRoll = "reRoll",
  alltoFalse = "alltoFalse",
  updateSorted = "updateSorted",
  addStatusEffects = "addStatusEffects",
  removeStatusEffects = "removeStatusEffects",
  addStatusEffect = "addStatusEffect",
  removeStatusEffect = "removeStatusEffect",
}